+++ novae/_constants.py
@@ -1,6 +1,3 @@
-import numpy as np
-
-
 class Keys:
     # obs keys
     LEAVES: str = "novae_leaves"
@@ -62,26 +59,4 @@ class Nums:
     # misc nums
     MEAN_NGH_TH_WARNING: float = 3.5
     N_OBS_THRESHOLD: int = 2_000_000  # above this number, lazy loading is used
-
-    def disable_lazy_loading(self):
-        """Disable lazy loading of subgraphs in the NovaeDataset."""
-        Nums.N_OBS_THRESHOLD = np.inf
-
-    def enable_lazy_loading(self, n_obs_threshold: int = 0):
-        """Enable lazy loading of subgraphs in the NovaeDataset.
-
-        Args:
-            n_obs_threshold: Lazy loading is used above this number of cells in an AnnData object.
-        """
-        Nums.N_OBS_THRESHOLD = n_obs_threshold
-
-    @property
-    def warmup_epochs(self):
-        return Nums.WARMUP_EPOCHS
-
-    @warmup_epochs.setter
-    def warmup_epochs(self, value: int):
-        Nums.WARMUP_EPOCHS = value
-
-
-settings = Nums()
+    RATIO_VALID_CELLS_TH: float = 0.7