+++ novae/utils/__init__.py
@@ -16,7 +16,7 @@ from ._utils import (
     sparse_std,
     iter_slides,
 )
-from ._validate import check_available_domains_key, prepare_adatas
+from ._validate import check_available_domains_key, prepare_adatas, check_has_spatial_adjancency
 from ._data import load_local_dataset, load_wandb_artifact, toy_dataset, load_dataset
 from ._mode import Mode
 from ._correct import batch_effect_correction