+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.1.1] - xxxx-xx-xx
+
+### Added
+
+- `novae.plot.connectivities` to show the cells neighbors
+- `novae.settings.auto_processing = False` to enforce using your own preprocessing
+
+### Fixed
+
+- Issue with `library_id` in `novae.plot.domains` (#8)
+
 ## [0.1.0] - 2024-09-11
 
 First official `novae` release. Preprint coming soon.